+++ ShadPractice/ShadPractice.Core/Models/Invoice.cs
@@ -10,9 +10,9 @@ namespace ShadPractice.Core.Models
     public class Invoice: Base
     {
         public int CustomerId { get; set; }
-        public double TotalItbis { get; set; }
-        public double SubTotal { get; set; }
-        public double Total { get; set; }
+        public decimal TotalItbis { get; set; }
+        public decimal SubTotal { get; set; }
+        public decimal Total { get; set; }
 
         [ForeignKey("CustomerId")]
         public virtual Customer Customers { get; set; }
