+++ ShadPractice/ShadPractice.Core/Models/InvoiceDetail.cs
@@ -12,10 +12,10 @@ namespace ShadPractice.Core.Models
         public int InvoiceId { get; set; }
         public int CustomerId { get; set; }
         public int Qty { get; set; }
-        public double Price { get; set; }
-        public double TotalItbis { get; set; }
-        public double SubTotal { get; set; }
-        public double Total { get; set; }
+        public decimal Price { get; set; }
+        public decimal TotalItbis { get; set; }
+        public decimal SubTotal { get; set; }
+        public decimal Total { get; set; }
 
         [ForeignKey("InvoiceId")]
         public virtual Invoice Invoice { get; set; }
